confirm for edi quote backend

+++ API/cotas.php
@@ -15,26 +15,26 @@ if ($_SERVER["REQUEST_METHOD"] == 'GET') {
 
         //Selección de id (ROGER)   
         if (isset($_GET["N"])) {
-    
+
             //VALIDACIÓN DE USUARIO
-    
+
             $sql_cota = mysqli_query($conexion_bd, "SELECT * FROM cota WHERE N=" . $_GET["N"]);
-    
+
             if (mysqli_num_rows($sql_cota) > 0) {
-    
+
                 $cota = mysqli_fetch_all($sql_cota, MYSQLI_ASSOC);
                 echo json_encode($cota);
             } else {
-    
+
                 echo json_encode(["success" => 0]);
             }
-    
+
             exit();
         }
     } else {
-    
+
         echo json_encode(["error" => "No estás autorizado", "code" => 401]);
-    
+
         exit();
     }
 }
@@ -49,6 +49,28 @@ if ($_SERVER["REQUEST_METHOD"] == 'PUT') {
             echo json_encode(["error" => "El campo clave no puede esta vacío"]);
         } else {
             validateFields($data);
+            $sql_cota = mysqli_query($conexion_bd, "SELECT * FROM cota WHERE N=" . $data->N);
+
+            if (mysqli_num_rows($sql_cota) > 0) {
+
+                $cota = mysqli_fetch_all($sql_cota, MYSQLI_ASSOC);
+                $cota = json_decode(json_encode($cota[0]));
+                //VERIFICAMOS QUE NO EXISTA UN LIBRO CON ESA COTA
+                $sql_libros = mysqli_query($conexion_bd, "SELECT * FROM libros WHERE cota = '" . $cota->cota . "' AND cod_isbn = '" . $cota->cod_isbn . "' AND cutter = '" . $cota->cutter . "' AND cota_completa = '" . $cota->cota_completa . "'");
+
+
+                if (mysqli_num_rows($sql_libros) > 0) {
+
+                    $libros = mysqli_fetch_all($sql_libros, MYSQLI_ASSOC);
+
+                    mysqli_query($conexion_bd, "UPDATE libros SET cota = '" . $data->cota . "', cota_completa = '" . $data->cota_completa . "', cutter = '" . $data->cutter . "', cod_isbn = '" . $data->cod_isbn . "' WHERE cota = '" . $cota->cota . "' AND cod_isbn = '" . $cota->cod_isbn . "' AND cutter = '" . $cota->cutter . "' AND cota_completa = '" . $cota->cota_completa . "'");
+                }
+            } else {
+                $cota = [];
+                $sql_libros = [];
+            }
+
+
             //Limpieza de datos a almacenar MODIFICADOS
             $data->cota = secureData($data->cota);
             $data->cod_isbn = secureData($data->cod_isbn);
@@ -57,8 +79,9 @@ if ($_SERVER["REQUEST_METHOD"] == 'PUT') {
             $data->ejemplar = secureData($data->ejemplar);
             $data->fecha_ing = secureData($data->fecha_ing);
             $data->cota_completa = secureData($data->cota_completa);
-    
-    
+            $libros = mysqli_fetch_all($sql_libros, MYSQLI_ASSOC);
+
+
             mysqli_query($conexion_bd, "UPDATE cota SET cod_isbn ='" . $data->cod_isbn . "', cota ='" . $data->cota . "', cutter ='" . $data->cutter . "', volumen ='" . $data->volumen . "', ejemplar ='" . $data->ejemplar . "', fecha_ing ='" . $data->fecha_ing . "', cota_completa ='" . $data->cota_completa . "' WHERE N ='" . $data->N . "'");
             echo json_encode([
                 "success" => "La cota se ha editado exitosamente",
@@ -73,14 +96,12 @@ if ($_SERVER["REQUEST_METHOD"] == 'PUT') {
             ]);
         }
         exit();
-    
     } else {
-    
+
         echo json_encode(["error" => "No estás autorizado", "code" => 401]);
-    
+
         exit();
     }
-
 }
 
 //DELETE ELIMINAR
@@ -95,17 +116,34 @@ if ($_SERVER["REQUEST_METHOD"] == 'DELETE') {
 
             echo json_encode(["error" => "los campos no pueden estar vacíos"]);
         } else {
-    
-            mysqli_query($conexion_bd, "DELETE FROM cota WHERE N='" . $data->N. "'");
-    
-            echo json_encode(["success" => "La cota fue eliminada de forma exitosa"]);
+
+            //VERIFICAMOS QUE NO EXISTA UN LIBRO CON ESA COTA
+            $sql_libros = mysqli_query($conexion_bd, "SELECT * FROM libros WHERE cota='" . $data->cota . "' AND cod_isbn='" . $data->cod_isbn . "' AND cutter='" . $data->cutter . "' AND cota_completa='" . $data->cota_completa . "'");
+
+
+
+            if (mysqli_num_rows($sql_libros) > 0) {
+
+                if (!isset($data->confirmed) || !$data->confirmed) {
+
+                    $libros = mysqli_fetch_all($sql_libros, MYSQLI_ASSOC);
+                    $foundBookTitle = $libros[0]["titulo"];
+                    echo json_encode(["bookFound" => true, "foundMessage" => "Esta cota ya tiene un libro asignado ($foundBookTitle). ¿Estás seguro de que deseas eliminarla? El libro no tendrá una cota asignada."]);
+                    exit();
+                } else {
+                    $sql_libros = mysqli_query($conexion_bd, "UPDATE libros SET cota = '', cota_completa = '', cutter = '', cod_isbn = '' WHERE cota = '" . $data->cota . "' AND cod_isbn = '" . $data->cod_isbn . "' AND cutter = '" . $data->cutter . "' AND cota_completa = '" . $data->cota_completa . "'");
+                }
+            }
+
+            mysqli_query($conexion_bd, "DELETE FROM cota WHERE N='" . $data->N . "'");
+
+            echo json_encode(["success" => "La cota fue eliminada exitosamente"]);
         }
         exit();
-    
     } else {
-    
+
         echo json_encode(["error" => "No estás autorizado", "code" => 401]);
-    
+
         exit();
     }
 }
@@ -125,11 +163,10 @@ if ($_SERVER["REQUEST_METHOD"] == 'POST') {
         echo json_encode(["success" => "La cota se ha registrado exitosamente"]);
 
         exit();
-    
     } else {
-    
+
         echo json_encode(["error" => "No estás autorizado", "code" => 401]);
-    
+
         exit();
     }
 }
@@ -138,37 +175,35 @@ if ($_SERVER["REQUEST_METHOD"] == 'POST') {
 $data = json_decode(json_encode($_GET));
 if (isAuthorized($data, $conexion_bd, true, true)) {
     $query = filtrarBusqueda($_GET, 'cota');
-    
+
     $sql_cota = mysqli_query($conexion_bd, $query);
-    
+
     $resData = paginar($sql_cota, $_GET, 'cota');
-    
+
     $sql_cota = mysqli_query($conexion_bd, $resData["query"]);
-    
-    
+
+
     //GUARDAR LA CANTIDAD DE FILAS EN UNA VARIABLE PARA FACILIDAD DE USO
     $rows = mysqli_num_rows($sql_cota);
-    
+
     $resData["pagination"]["end"] = $resData["pagination"]["start"] - 1 + $rows;
-    
+
     if ($rows > 0) {
-    
+
         $cota["data"] = mysqli_fetch_all($sql_cota, MYSQLI_ASSOC);
-    
-    
+
+
         //AGREGAR INFORMACION UTIL PARA EL FRONTEND
-    
+
         $cota["pagination"] = $resData["pagination"];
-    
+
         echo json_encode($cota);
     } else {
-    
+
         echo json_encode(["success" => 0]);
     }
-    
-    exit();
 
-    
+    exit();
 } else {
 
     echo json_encode(["error" => "No estás autorizado", "code" => 401]);
